cleaning and slightly refactoring front-end
