+++ README.md
@@ -161,64 +161,5 @@ docker compose -f compose-prod.yaml logs
 docker compose -f compose-prod.yaml exec <service_name> <command>
 ```
 
-Or, in a bash profile like file, we can specify `export COMPOSE_FILE=compose-prod.yaml`
-
-## Used Crates
-
-### FrontEnd
-
-To format the component code in the `view!{ ... }` macros,
-
-```bash
-cargo install leptosfmt
-```
-
-- `cargo add leptos@0.6 --features=csr`
-  - Required for using Leptos as our frontend framework.
-- `cargo add leptos_router --features=csr`
-  - Since this is SPA, we want to give illusion of routing with a router.
-- `cargo add serde@1.0 --features=derive`
-  - Required for serializing and deserialization
-- `cargo add serde_json@1.0`
-  - Required for serializing and deserialization
-- `cargo add wasm-bindgen@0.2`
-  - This is crate of WASM bindings to JavaScript web APIs, needed to fetch.
-- `cargo add wasm-bindgen-futures@0.4`
-  - This is crate of WASM bindings to JavaScript web APIs, needed Promises and futures (for fetching).
-
-### BackEnd
-
-- `cargo add actix-web@4.6`
-  - Required for using Actix-Web as the backend framework.
-- `cargo add tokio@1.37 --features=macros,rt-multi-thread`
-  - Following Zero to Production, Tokio is an asynchronous runtime for Rust.
-- `cargo add actix-cors@0.7`
-  - Cannot get separate frontend without it, see [Cors docs](https://docs.rs/actix-cors/latest/actix_cors/)
-- `cargo add tracing@0.1 --features=log`
-  - Better logs for asynchronous applications
-- `cargo add tracing-subscriber@0.3 --features=registry,env-filter`
-  - To help implement and use the `tracing::Subscriber` trait
-- `cargo add tracing-bunyan-formatter@0.3`
-  - Handy crate to format logs as JSON
-- `cargo add tracing-log@0.2`
-  - When actix-web fires a log event, this crate can redirect logs to our tracing subscriber.
-- `cargo add tracing-actix-web@0.7`
-  - Provides `TracingLogger` to be used as middleware to collect telemetry data.
-- `cargo add thiserror@1`
-  - Provides procedural macro to derive `std::error::Error` trait.
-- `cargo add anyhow@1`
-  - Helps to simplify error handling by being like a catch-all error trait object.
-- `cargo add --no-default-features reqwest --features=json,rustls-tls,cookies`
-  - This is easy-to-use HTTP client, which is very helpful in testing
-- `cargo add --dev wiremock`
-  - Can spawn mock servers for testing.
-  - This will only be used for testing purposes, no need to spawn mock servers in live code.
-  - Actix-Web testing does not appear to necessarily spawn a server for integration tests.
-- `cargo add serde@1 --features=derive`
-  - Rust standard for data serialization and deserialization.
-- `cargo add serde-aux@4.5`
-  - Houses helpful function for casting types during deserialization.
-- `cargo add serde_json`
-  - Providing serialization and deserialization implementation for JSON format.
-- `cargo add config`
-  - Allows you to read and merge configuration from multiple sources.
+Or, in a bash profile like file, we can specify `export COMPOSE_FILE=compose-prod.yaml`.
+I only have one configuration so need to uncomment code for production in nginx.conf.
