+++ frontend/Dockerfile.dev
@@ -9,5 +9,4 @@ ENV TRUNK_SERVE_ADDRESS 0.0.0.0
 ENV TRUNK_SERVE_PORT 8080
 WORKDIR /app
 COPY . .
-# CMD ["trunk", "serve", "--port", "8080"]
 CMD ["trunk", "serve"]
