+++ frontend/nginx.conf
@@ -8,15 +8,13 @@ http {
   # -- removes NGINX Version from footer of errors
   server_tokens off;
   charset utf-8;
-  # include /etc/nginx/mime.types;
-  # default_type application/octet-stream;
 
   server {
     # default_server is just a flag, no value
     listen 80 default_server;
 
     server_name quiztestapp.io www.quiztestapp.io;
-    # server_name kevsquizapp.com;
+    # server_name kevsquizapp.com www.kevsquizapp.com;
 
     # -- Renewal Challenge
     location ^~ /.well-known/acme-challenge/ {
