+++ frontend/src/components/card.rs
@@ -2,6 +2,10 @@
 //! This is a simple component to hold other components
 use leptos::*;
 
+// TODO: Since other card types have been included, this should be renamed to something more
+// specific.
+/// Encapsulates children components in a 'card' to provide certain consistent styling
+/// througout application.
 #[component]
 pub fn Card(children: Children, on_click: Option<Callback<ev::MouseEvent>>) -> impl IntoView {
     view! {
@@ -19,6 +23,9 @@ pub fn Card(children: Children, on_click: Option<Callback<ev::MouseEvent>>) -> i
     }
 }
 
+/// Component to indicate that certain buttons and other components added to display
+/// are not yet ready for use. It allows for application to contain unimplemented ideas
+/// without leading user's to think a button is not working as intended.
 #[component]
 pub fn TodoCard(children: Children, on_click: Option<Callback<ev::MouseEvent>>) -> impl IntoView {
     view! {
@@ -39,6 +46,7 @@ pub fn TodoCard(children: Children, on_click: Option<Callback<ev::MouseEvent>>) 
     }
 }
 
+/// For enclosing children components in a centered component.
 #[component]
 pub fn CenterFormCard(children: Children) -> impl IntoView {
     view! {
