+++ frontend/src/components/dashboard/get_quiz.rs
@@ -2,22 +2,14 @@
 //! This component will handle quiz making logic and pass
 //! user to the making questions screen.
 use leptos::*;
-use serde::{Deserialize, Serialize};
 use web_sys::{Headers, RequestMode, Response};
 
 use crate::{
-    components::Card,
-    models::mimic_surreal::{SurrealQuiz, Thing},
-    store::AppSettings,
-    utils::Fetcher,
+    components::Card, models::mimic_surreal::SurrealQuiz, store::AppSettings, utils::Fetcher,
 };
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct QuizJsonPkg {
-    pub name: String,
-    pub description: String,
-}
-
+/// The Show Case is like a container to display all quizzes by a user.
+/// Most / All props passed into this component are for prop-drilling.
 #[component]
 pub fn QuizShowCase(
     quiz_list: RwSignal<Vec<SurrealQuiz>>,
@@ -26,10 +18,8 @@ pub fn QuizShowCase(
     quiz_updater: Callback<SurrealQuiz>,
     quest_calibrate: Callback<SurrealQuiz>,
 ) -> impl IntoView {
-    //  -- Create Signals --
-    //  -- Create References --
-    let quiz_title: NodeRef<html::Input> = create_node_ref();
-    let quiz_description: NodeRef<html::Textarea> = create_node_ref();
+    // -- Create Signals --
+    // -- Create References --
     // -- Use Context --
 
     // -- Render View --
@@ -57,6 +47,8 @@ pub fn QuizShowCase(
     }
 }
 
+/// Many Quiz Exhibits live inside the Quiz Showcase.
+/// Component manages display of individual quizzes.
 #[component]
 pub fn QuizExhibit(
     surreal_quiz: SurrealQuiz,
@@ -105,6 +97,7 @@ pub fn QuizExhibit(
         }
     });
 
+    // -- Render View --
     view! {
         <Card on_click=None>
             <h3>"Name: "{move || quiz_sig.get().name}</h3>
